requirements.txt
appnope==0.1.0
decorator==4.0.4
Django==1.8.4
django-filter==0.11.0
django-rest-swagger==0.3.4
djangorestframework==3.2.4
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
PyYAML==3.11
simplegeneric==0.8.1
traitlets==4.0.0
wheel==0.24.0
